Added my own wordnik api key and nounlist fallback
Finally got my wordnik api key, just replaced the temporary one. I am gonna leave it in as this is just a fun little couple day project and I probably won't use the wordnik api in the future. I guess, don't abuse the key, but use it if you really need to!