Added babel to transpile es6.
Figured out that the reason the site didn't work on IOS 10.3 and lower was because I was using es6 keywords const and let, Safari Mobile 10.3 and lower didn't know what those were and so didn't execute anymore of the code. Babel will change all of them to var inside of the main.min.js file only.